Added points system
Added a points system that allows towers, projectiles, and the player projectile to increment the points of each individual  player meaning that each player has their own separate points that get incremented on the server side and broad cast back to the clients that need their points adjusting. Also added improved ballista functionaility meaning that it only targets enemies that are alive and not in their death animation.