Added more robust functionaility to help prevent leaderboard desyncs
changed the on rep notify for the leaderboard state in the server object so that it catches when the values of already existing players are updated  and marked for replication to all other server object proxies on the clients. This replcaces a multicast that was previously doen to update the leaderboard state on the main charcter itself which was inconsistent due to the volume of multicasts that could be produced at one time when attempting to update the leaderboard state on the main character object itself.